Use proactive updates to refresh nodes user-data
Allow instance groups to proactively update nodes in "REFRESH" mode only. That
will mean that group managers can only attempt nodes' config in place and not
start/stop instances. Since we are allowed to patch user-data that way we can
use a daemon like kured to roll ignition updates to the cluster.